refactor to break circular dependency
